+++ apps/dashboard/app/javascript/application.js
@@ -56,6 +56,6 @@ jQuery(function(){
     title: function(){ return $(this).text() }
   });
 
-  $('[data-toggle="popover"]').popover();
-  $('[data-toggle="tooltip"]').tooltip();
+  $('[data-bs-toggle="popover"]').popover();
+  $('[data-bs-toggle="tooltip"]').tooltip();
 });