+++ apps/dashboard/app/javascript/path_selector/path_selector_data_table.js
@@ -1,3 +1,4 @@
+import { alert } from '../alert';
 
 export class PathSelectorTable {
   _table = null;
@@ -12,6 +13,7 @@ export class PathSelectorTable {
   modalId             = undefined;
   showHidden          = undefined;
   showFiles           = undefined;
+  filePattern         = undefined;
 
   constructor(options) {
       this.tableId             = options.tableId;
@@ -23,6 +25,7 @@ export class PathSelectorTable {
       this.modalId             = options.modalId;
       this.showHidden          = options.showHidden === 'true';
       this.showFiles           = options.showFiles === 'true';
+      this.filePattern         = options.filePattern;
 
       this.initDataTable();
       this.reloadTable(this.initialUrl());
@@ -84,7 +87,7 @@ export class PathSelectorTable {
   async reloadTable(url) {
     try {
       $(this.tableWrapper()).hide();
-      $('#loading-icon').show();
+      $(this).closest('.loading-icon').show();
       const response = await fetch(url, { headers: { 'Accept': 'application/json' }, cache: 'no-store' });
       const data = await this.dataFromJsonResponse(response);
       $(`#${this.breadcrumbId}`).html(data.path_selector_breadcrumbs_html);
@@ -104,7 +107,7 @@ export class PathSelectorTable {
   }
 
   resetTable() {
-    $('#loading-icon').hide();
+    $(this).closest(".loading-icon").hide();
     $(this.tableWrapper()).show();
     $('#forbidden-warning').addClass('d-none');
   }
@@ -207,6 +210,16 @@ export class PathSelectorTable {
 
   // filter the response from the files API to remove things like hidden files/directories
   filterFileResponse(data) {
+    let regex = undefined
+
+    try {
+      if (this.filePattern !== undefined) {
+        regex = RegExp(this.filePattern);
+      }
+    } catch {
+      alert("The regular expression provided for this path selector did not compile");
+    }
+
     const filteredFiles = data.files.filter((file) => {
       const isHidden = file.name.startsWith('.');
       const isFile = file.type == "f";
@@ -216,7 +229,7 @@ export class PathSelectorTable {
       } else if(isHidden) {
         return this.showHidden;
       } else if(isFile) {
-        return this.showFiles;
+        return this.filteredByFilename(file, regex);
       } else {
         return true;
       }
@@ -224,5 +237,18 @@ export class PathSelectorTable {
 
     data.files = filteredFiles;
     return data;
+  }
+
+  filteredByFilename(file, regex) {
+    if (regex !== undefined) {
+      if (file.name.match(regex)) {
+        return this.showFiles;
+      } else {
+        return false;
+      }
+    }
+    else {
+      return this.showFiles;
+    }
   }
 }