+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [4.0.0] - 12-30-2024
+
 ### Added
 - BatchConnect form labels can now be made dynamic with data-label-* in [3598](https://github.com/OSC/ondemand/pull/3598).
 - BatchConnect form auto_modules directive can now filter by string or regex in [3574](https://github.com/OSC/ondemand/pull/3574).
@@ -30,6 +32,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - UIDs can now be returned by the mapper script in [3795](https://github.com/OSC/ondemand/pull/3795).
 - XDMoD jobs widget now shows CPU, Memory and walltime in [3789](https://github.com/OSC/ondemand/pull/3789).
 - Global batch connect form items can now be defined in ondemand.d files in [3840](https://github.com/OSC/ondemand/pull/3840).
+- The path_selector widget now supports filtering results in [3992](https://github.com/OSC/ondemand/pull/3992).
+- OOD now responds to `/etc/ood/config/plugins` to support eaiser customizations in
+  [3991](https://github.com/OSC/ondemand/pull/3991) and [4020](https://github.com/OSC/ondemand/pull/4020).
+- Dynamic batch connect support for data-exclusive-option-for in [4025](https://github.com/OSC/ondemand/pull/4025).
+- Better localization support in [4003](https://github.com/OSC/ondemand/pull/4003).
 
 ### Changed
 - Script models have been renamed to Launcher in [3397](https://github.com/OSC/ondemand/pull/3397).
@@ -67,6 +74,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - OnDemand packages no longer relies on scl at runtime in [3952](https://github.com/OSC/ondemand/pull/3952).
 - Only root owned ondemand.d files will be loaded in prodution in [3969](https://github.com/OSC/ondemand/pull/3969).
 - The files app will now create a spinner on the files table when making new requests in [3973](https://github.com/OSC/ondemand/pull/3973).
+- nginx_clean utiltiy in nginx_stage will now determine inactive users and clean their PUNs too in [3942](https://github.com/OSC/ondemand/pull/3942).
 
 ### Fixed
 - Ensure that the asset directory is clean when building in [3356](https://github.com/OSC/ondemand/pull/3356).
@@ -92,6 +100,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Batch jobs now specify workdir, fixing issues with submit_host jobs in [3913](https://github.com/OSC/ondemand/pull/3913).
 - Javascript that queires for atch connect sessions will create an alert div and stop polling if it fails in [3915](https://github.com/OSC/ondemand/pull/3915).
 - auto_qos correctly returns one option for each qos in [3955](https://github.com/OSC/ondemand/pull/3955).
+- The download button will now disable when users have selected non-downloadable files in [4008](https://github.com/OSC/ondemand/pull/4008)
+- Fixed a bug where the icon shows up on 2nd path_selector in [4009](https://github.com/OSC/ondemand/pull/4009).
+- nginx_stage no longer caclulates the users' groups in [#4012](https://github.com/OSC/ondemand/pull/4012).
 
 ### Security
 
@@ -1475,7 +1486,8 @@ Similar changelog as [3.0.0]. This version was not released to the general publi
 ### Changed
 - From 1.3.7 - 1.4.2 updated app versions
 
-[Unreleased]: https://github.com/OSC/ondemand/compare/v3.1.0...HEAD
+[Unreleased]: https://github.com/OSC/ondemand/compare/v4.0.0...HEAD
+[4.0.0]: https://github.com/OSC/ondemand/compare/v3.1.10...v4.0.0
 [3.1.10]: https://github.com/OSC/ondemand/compare/v3.1.9...v3.1.10
 [3.1.9]: https://github.com/OSC/ondemand/compare/v3.1.7...v3.1.9
 [3.1.7]: https://github.com/OSC/ondemand/compare/v3.1.4...v3.1.7